nested_schemas/api_requirements.txt
# Flask API Server Requirements
# For exposing Python tools as REST APIs

Flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1
Werkzeug==3.0.1